requirements/dev.txt
trafaret==1.2.0
trafaret_config==2.0.2

fastapi==0.59.0
uvicorn==0.11.5

aiohttp==3.6.2
aiofiles==0.5.0

orjson==3.2.1

PyJWT==1.7.1
passlib==1.7.2
bcrypt==3.1.7
python-multipart==0.0.5

asyncpg==0.20.1
gino[starlette]==1.0.1

PyYAML==5.3.1
